Upgrade from Microsoft.Azure.Cosmos.Table to Azure.Data.Table.
** Backup/Restore is NOT compatible with prior versions **

Upgrade from Microsoft.Azure.Cosmos.Table to Azure.Data.Table.   The Microsoft.Azure.Cosmos.Table library was deprecated on 3/31/2023.  The Azure.Data.Table API supports serialization directly and has different typing of serialized data than achived through the customized DynamicTableEntityJsonConverter/Serializer.  As a result, backup files created with prior versions are not compatible to be restored with this release.  Similarly, backups taken with this release may not be restored under prior versions.  It is possible to convert a prior file version to work with the new serialization.  At this time, no tool is being offered to do so.